+++ api/src/app.module.ts
@@ -6,6 +6,7 @@ import { CatalogModule } from './catalog/catalog.module';
 import { ClientsModule } from './clients/clients.module';
 import { InvoicesModule } from './invoices/invoices.module';
 import { OrdersModule } from './orders/orders.module';
+import { PaymentsModule } from './payments/payments.module';
 import { VendorsModule } from './vendors/vendors.module';
 
 @Module({
@@ -33,6 +34,7 @@ import { VendorsModule } from './vendors/vendors.module';
     ClientsModule,
     InvoicesModule,
     OrdersModule,
+    PaymentsModule,
     VendorsModule,
   ],
 })